Basic Interpretation & Fixture Worlds
This introduces the basic object for creating a lighting world
and fixtures, and apply certain parameters - probably in a naive
way.

The following concepts are created
- Selection Engine
- World
- Fixture

At the moment fixture params are defined by a DSL type syntax in the
class but eventually will move towards loading fixture profiles
that users define, and generating a class from that.

Worlds are created in a successive parent-child chain as they are
selected upon, which should allow for nested selections.

The selection engine is extremely naive. It currently only supports
single selections by id. However there are a huge number of features
to support here for a query syntax. At the work for selecting fixtures
happens inside the world, but this probably is a mistake? Future work
will reveal better delianation of responsibility.