Select Range of Fixtures and apply to them
Support is added for a new AST Node that defines an inclusive range
for selecting multiple fixtures at once.

A QueryBuilder is introduced which is a starting point for the more
complex format that the selection engine will need to understand. At
the moment it iterates a set of maps with associated metadata
and slowly builds up a world with the correct fixtures selected.