A Selection can have a Sub Selection
Because selections are just sub sets of their parent worlds, by
adjusting the grammer it's possible to add a check if we are making
another selection. This creates a new world with just those fixtures
inside.

Deselection is handled nicely as the tree walker leaves the branch,
returning to parent worlds as it walks back up.