Set the paramater to a descendent of Value rather than literal
This paves the way for fades in the future. In traditional terms,
when a cue/executor is evaluated, we want to the world state to
represent all the information needed to run the cue.

This prevents some interactions that could be possible for mutatable
state I think, but this should be fine...