@fade directive to fade to a value over time
This is the first directive of a more fully featured fading directive
support, but introduces the Time keywords and TimeBlock into the AST.

A Fade is generated at the point of application, if it is needed, but
results in a StaticValue otherwise.

Time is tracked via a linked set of TimeContexts. When a selection is
made, it inherits from it's parent block the TimeContext. In this way
it's possible to nest contexts etc.

When applying time information, the linked set of TimeContexts is walked
back up until a nil value is found. This will allow @snap overrides in
the future.