Expand Fixture DSL to accept group parameters
In preperation for position and colour information for fixtures we
need a much more comprehensive DSL for defining these.

Pan and tilt can be viewed as a related set of parameters under the
group of position, which will act as a virtual parameter. This commit
enables this by introducing new DSL syntax to do so.

A Parameter is a object that contains the definition of a parameter,
and eventually will hold such values as min, max, and dmx offset - as
well as handling DMX conversion lambdas / however the proposed solution
works.

A ParameterInstance is a thin wrapper over this which has the actual
value stored in it. The reason for this abstraction is to allow for
smart default parameters to be used for commonalities between fixtures
:)