Apply Anonymous Tuple to a virtual group parameter
A Tuple AST node is added which generates the correct anonymous tuple,
which is a hash where all keys are _N. This is then converted into
a ValueTuple in the interpreter.

Most of this abstraction work is being used to allow for ranges etc.
to occour between tuples and inside them. Their is some inherient
complexity to this, but once wrangled this should cover almost all the
application code - bar variable resolution which shouldn't be too
hard.