'_' defers to current value in application
By creating a nil Value in the AST we can tell the application not to
change the value, in essence creating a current value.

This passing around of nil feels a little dangerous, as we have to check
if nil is present a lot, but this little bit of a magic is worth the
price to pay for the ease of this solution. It avoids tangling world
state up with evaluated values, which is a big win :)