Function rgb() returns a color tuple for application
Functions are, like CSS, defined in code not by the user. But this
doesn't mean it won't let users have their own code space in the future
:)

FunctionRegister is the central dispatch for all functions, which are
called from the identifier in the code and it returns the procedure to
call with args. I think this brings the code gen almost to completion,
with only variables (or pallettes in traditional terms) left to add.

The parser does not follow a clean chain of precedence, which is
something I'd like to refactor out at some point...