Define and recall palletes of lighting information
Two new AST nodes are introduced, VarDefine and VarFetch, which act as
you can imagine.

VarFetch though will depend on it's context. When used as a value, will
result in a partial applicaton for that paticular parameter, and do
nothing if no information is found. Otherwise the global applicaiton
just executes the pallete in scope. This should allow pallete
information in functions, we will see :)