Range between Reference Values
The original range implementation assumed value literals for the start
and finish point of the range. However, in our case we want to be a
little more complex to allow ranging between pallettes.

There is a lot of corner cases to consider here, and for now we bail out
of most of them by saying if there are multiple values, we will use the
first one from start and end to create the range. However you could go
bananas with all sorts of ranging logic.. maybe in the future if I find
a need for it ;)