Basic SACN Implementation
This library provides the very rudimentary implementation of the SACN
data packet. It performs well, managed 6HZ with 2000 universes of
constant change data. In more realistic scenarios it should maintain at
least 20Hz. If speed becomes an issue then I already have eyed rewriting
the project in crystal :)

With this infrastructure in place it should now be possible to convert
the lighting world state into real lighting information. Which will be
the next major milestone for this project.