DMX Offsets in the Fixture DSL
The DSL now allows you to pass the dmx offset, and by extension I added
the default hash param as well for prosperity.

When the offset is omitted it will use a basic heurestic of one more
than the last one it saw to assign it. This cleans up the DSL for the
most common case, where params are defined in order.