Basic Universe Data Generation
The LightingEngine takes a World from the Interpreter and calls run on
all it's fixtures, which in turn calls run on all it's parameters,
passing the current engine time down the chain. This time will
eventually be used for dynamic effects but for now does nothing.

Every value type must implement run, which will generate a 0-100 percent
value based on the current time, or in the case of static value just
return it's value. The simple case :)

A Universe is a 512 long array of bytes, which the resulting footprint
of the Fixture is then pasted in to it. Included here is debug dump
output.

As Universes and Addresses are 1 indexed, there is the potential for off
by one errors everywhere :)