Refactor out data broadcast to Output class
Implemented SACNOutput class that will broadcast given universes as
SACN packets. This is a very simple implementation with no real
consideration yet for multicast et al. We're just trying to see some
output. In the future there will be more outputs, that output specific
universes in different ways, but let's keep it simple :)