Fades & Delays run to Time
By introducing a Timer class we can track the elapsed time since
running, and also keep track of how long it took to process this
paticular frame so we can get a consistent FPS under high work loads.

A Fade has relatively simple logic. But for a delay we must ensure that
the time that the final value receives has been offset by the delay
amount. Imagine a 3s delay and a 3s fade. Without this offset after the
delay expired the fade would immediately complete, which is not the
desired behaviour.