Refactor group parameters to handle their own tuples
Before group parameters were ghost parameters, a virtual parameter that
only handled reference to what its children were. Now it looks after its
own children and has its own assigned tuple as a value.

This change was motivated by the fact that functions will want to return
their own tuples. For example, what I was trying to implement was a
colour cycle. Now that is theoretically possible.

This is a wide ranging refactor, that ends up with parameter application
code out of the Fixture class which is good. I am sure that I will find
many many corner cases I have yet to handle. One functionality that has
been lost is talking to individual parameters e.g pan: and tilt: of the
group. I have to think about how much I care about that.. seeing as the
(_, 10) is the same as tilt: 10.