Live reload world state when watched file changes
This isn't the cleanest implementation of this idea, because at the
moment we are doing a pretty hack and slash runtime environment in
general.

When we implement real show files we can think about what exactly is the
lux 'core' which will be seperate from the runtime gui etc :)

We are able to mutate the state in the file listener, updating the world
while retaining the current cue time, so that the fades etc. don't
rerun. There are scenarios when we should rerun, and we will handle them
later.. let's keep it basic for now.