Chase Fixtures
To create an individual fixture chase was a challenging at first, but
when you think in the application of waves becomes a little easier. This
avoids having to rerun the AST over and over. Now we've done this once,
it's easier to think in terms of wave construction.

Also included is a square wave.

DynamicValues that return DynamicValues will run again until they
resolve to some real value.