Resolve '_' in DynamicValues
When a dynamic value is applied, it may have a nil argument, which we
will subsitute for the current parameter value. The only time a user
should use a '_' is when referring to an argument that is setting a
value. Though there is nothing stopping them from doing it in another
argument, it's just that's not what they are expecting.

It would be theoretically possible to detect these changes, for example
if you wanted to fade to a sin wave of a longer period, it would copy
over arguments from the fade before and then resolve and remaining nils.

I'll implement that if I ever find a need for it :)