Folder/File Based CueLists
A CueEngine handles a set of CueLists which are identified by their
folder names in a cues/ folder. A CueLists has no concept of a cue,
because a cue is simply a file to execute.

We are able to load a cue list, which will enter the first sorted file
cue. And we can then go, which currently will loop around a cuelist at
the end, and run the cue.

We detect changes in files in the cue list, and ask the CueEngine for
the files we need to rerun to build the world state after this change.
This could be a somewhat expensive operation - especially when we
consider other cuelists that could be running. However it only affects
programming - playback should still be a lot cheaper.

Included is a simple cue list which I can play with :)