Goto to any cue and load correct state
When going forwards to a cue we want to execute each cue in turn so that
we get the right state. When going backwards, we have to rebuild from
the empty world.

When looping around via go from the last cue to the start, we don't
currently clear the world and start again. Some cuelists you wouldn't
want this to happen - e.g effects. But on the loaded cue list maybe you
do? A 'reset' keyword may be the answer