Refactor cue paths to a Cue Object
Becasue we want to be able to do some additional things with cues, we
want to be able to store information on them - thus converting them to a
Cue object.

We are mainly passing file paths around in the interpreter. This maybe
doesn't feel that clean but it is better than passing cue objects
because not every evaluation etc. is a cue. Though granted there is some
code cleaning to be done here - however I increasingly look forward to a
rewrite in a fast type safe language, with tests, rather than keeping to
add to this ball of mud prototype - now that the domain is well explored
:)