Basic Fixture Library from .yml Files
Rather than defining our own format from scratch we can use the
expressive .yml format for defining fixture profiles. This is an easy to
write, human readable format which easily maps to ruby code definitions.

This commit is the very barebones of this infrastructure, recursively
loading the fixture library folders and creating anonymous classes with
the correct DSL functions called.