Fixture Profile Support - Robe 1200 AT (MODE 1)
This profile is semi-complete, in the sense that it correctly represents
the mapping of values to the DMX mode 1 - but it does not yet support
the listed 'locals' for the representation of DMX values for color
wheels etc. - coming soon :)

The calculation of current_offset is now much simpler and reflects
expected 1 based indexing of most DMX charts.

A paramter now be defined as fine, which will generate it's values as
two DMX values, a main part and a fractional part. This is most often
the case for pan/tilt in fixtures.