PercentValue returns the static numeric between min and max range
When run, percent value returns the right numeric for it's current
place. To do this we need to tell it what parameter it's applied to at
run time. This may become an issue if you try to fade from a percent
value to a static value as well - as you want to marshal them to the
same type before calculating the fade..