Addable Address Behaviour
A address is the combination of a universe number and an address,
following the common nomenclature - e.g 1.001, or 3.051.

Adding addresses together is useful for working out sequential patch
numbers. But in the case that we exceed 512, it should wrap to the next
start address in the next universe, which is invariable 1. For example:

1.510 + 0.100 = 2.001

Supported is the addition of unvierses, but this doesn't make a huge
amount of sense in actual application.