Major Refactor of Value / Generator Structure
This mainly comes down to thinking in a way that plays better with a
data first approach rather than an object first approach. Due to
limitations from trait objects.

All valid values fall under an enum, that contains their matching
struct. We introduce a Visitor pattern for conversion that currently is
incomplete but will be fleshed out more.

Overall the code is in fact simpler. And more importantly allows the
right abstraction for containing multiple generators of multiple
different values.