Apply DmxStrings of ResolvedFixtures to the Multiverse
The Multiverse is the collection of all output Universes, and are
created at the point of application of a given DmxString to an Address.
It is possible that a Universe is no longer required, and could be
dropped - but as they are only 512 bytes this may be something to worry
about if it evers becomes an issue.