Fade has a relative start time
First time a fade is generated, it captures the current elapsed time,
and calculates it's own time as an offset from there.

This implementation may, at current, be naive. I'd like the whole
lighting engine to be timecode native, so it would calculate it's offset
based on the timecode at point of triggering. I want to be able to scrub
through time and always evaluate to the same lighting state.