Build a Query and mutate Fixtures based on it
The QueryBuilder provides a convenient interface for creating a query
struct.

The Query is a ordered collection of steps to execute, which is then
evaluated step by step to build up a (for now unordered) set of fixture
id's.

These fixture id's are used to get a filtered mutable iterator over the
fixture set, which we can then do whatever we want over. The reason we
use ID's rather than the fixtures themselves, is because this is more
flexible re: borrow checking. It's basically an arena, and we can store
these ID sets anywhere after evaluation (groups, in selectors, etc.) In
the future we will have some mechanism for caching evaluation,
invalidating if patch changes - though patch changes will be disallowed
during runtime, and patch changes cause a system reload.

This is really a first cut of what will be one of the unique and
powerful features of the ecosystem, and represents the simplest possible
representation of how it _could_ work.