Incomplete timecode Time Struct with TypeState
This commit is a WIP of the timecode Time struct. We want to be able to
use the same time struct in a lot of places, but with various frame
rates. This approach uses type state to store that FrameRate.

I'm making this commit because I think this approach might be over
complex - frame number is the only differing calculation and could just
be derived from a stored nanosecond value on the time (as could the
other elements of the time frame!)

I'll sleep on it and come back to it.