Much simpler Time struct
This is a complete cut of the Time struct, which allows the stored
nanoseconds value to be converted correctly to hours, minutes, seconds
and frames.

We can not use the From trait to do the conversion from a duration,
because we need to pass in the FrameRate enum. I'm happy with this trade
off for the reduced complexity.