Fix strange time race condition in Timer
The original pause algorithim used the stored state of an Instant
to determine the pause. However this lead to wierd failign tests
that sometimes would or wouldn't pass.

This solution tracks the state of the timer, and captures at
the point of pausing the duration that has elapsed, and will
use that for output instead.