A Track has many Actions with many Selections (wip)
This is a Work In Progress Commit.

This many not be the right architecture, though is a first stap at
what it may look like. It doesn't work yet, and there are problems
with the borrow checker galore - probably treating environment
as an object too much rather than a data collection.

An Action has many Selections. A Selection is a combination of a
query over the fixture set, and applicators. An applicators is  a
bad name for a pair of Param and Generator.

This is just a hashing out of ideas, and is sure to be replaced by
something my brain comes up with in my sleep :)