Use Apply struct to pair a Param and a Generator
An Apply is the combination of a target Param and the generator
that should be assigned to it. We also have to implement traits to
allow for the copying of Generators, which for now means some
'static lifetimes, but might bite us in the future.

This is the first commit of a branch exploring a different route
toward tracks and actions.