ApplyGroup ties together a Query and many Apply's
Rarely to a group of fixtures do you want to only apply one param,
so this ties together a query and the applications that you want
to make to it. This concept comes from the original Lux language
prototype. And while isn't strictly necessary, models that domain
well - and this engine is closely related to the language.