Revert an Action by walking back up the history vector
This is the very bare basics of the action undo system, which is
implemented by cloning the FixtureSet into a history vector. We
can then revert by popping the last value and assinging it to the
environment.

To facilitate this we have to create our own implementions of
clone to deal with the cloning of dyn Generators and such :)