Tracks have time registered Actions that time travel
This algorithim, as of yet untested - but should because of it's
strangeness be tested to death soon - allows a track to have a
time ordered set of actions, that applies any unapplied actions
when time moves forward normally.

If the time has gone backwards, we revert to the latest history
we know about, and then rebuild any actions that aren't included
by that history.

There are some naive assumptions in this code regarding the
identification of histories, and the mutable ref to environment.
I don't like using output mutable references, but as of yet makes
the two places they are used easier to reason about.