Remove relationship of Time to FrameRate
It's actually much easier in API terms not to have a framerate
associated with a time. In fact, it might be unnecessary for the
Source to have a framerate - as it's virtual. In any case, we make
it so getting a frame number requires a FrameRate.

When we have a real timecode source, and it sends frame info, this
api will become a little more clear.