WIP: Refactor track action code to not need mut ref of the Env
This is a multi commit effort to refactor out the prototype spike
of code for the application of Actions in Tracks, and the time
travelling History.

We start by establishing a set of test cases to work towards, and
a host of refactoring work for a cleaner design. This means moving
out a lot of things to sub structs (work on going) and tidying up
interfaces.

Future commits will continue to work on test cases, tidy up API
design. Much work is to be done on code quality, but for now we
choose to accumulate technical debt to work towards getting core
systems functional. But it will be important to start this work
soon, before the codebase becomes unwieldy.