When time moves backwards restore the closest History
This is an attempt to fully test the logic for when time moves
backwards, and restoring a History of the Environment. However,
we can be certain that as the human mind only progresses in the
time foward direction, there are some corner cases that will need
correcting - it is the intention to build test cases every time a
bug is discovered.

The general algorithm is such:
1. Find the closest actions to the point in time across all tracks
   that have a history
2. From those, find the actual closest action
3. If there isn't one, restore state as if time is 0
4. If there is, restore that history, discard all other histories
   and tidy up histories of actions in all tracks after that point