Action's give their generators a start time when Applied
We need to create a unqiue copy of the action with a start time
applied to generators that need.

This is because say we have a long running fade, and we jump to
the middle of that state, we want to be able to calculate the
fade relative to that, where as before we assumed the first call
to generate the fade was the start time.