Lux: Build AST for simple assign statement
Though I have implemented several hand rolled parsers, I've
decided that it would be prudent to use a parser generator library
for Lux, to go faster and with more batteries included.

This first comment enables the basic syntax:

```
intensity: 100
```

Which equates to saying set the intensity of all selected fixtures
to the literal value 100.