Lux: Consolidate literal numbers to a Numeric AstNode
In lumen, we expect all the numeric values to be f64, so there is
no point distinguishing between them here. This assumption may
change in the future - if it becomes clumsy to assure that certain
numbers are whole numbers (e.g when talking about fixture ID's).

'10.0', '10', '5.5235' are all valid.
'10.' is not valid.