Lux: Parse Select fixtures in nested Select Blocks
A Select Block is a query, followed by a series of statements,
that limits the execution to only the fixtures that are valid for
that query. Nested select blocks further constrain the scope.

```
1 {
    intensity: 100
}
```

This is not a full implementation of the query grammer, which will
be quite complex and fully fleshed out a later date.