Lux: Evaluate a global Apply Statement
This is the very first commit at evaluating a syntax tree from the
evaluator into a Lumen Environment. To get a foothold in this
thorny problem, we start with just evaluating everything to a
global action, applied at time 0. This is so we can get a full
pipeline working as soon as possible - before adding more complex
features :)

Given a .lux file with

```
intensity: 100
```

We evaluate the correct static generator to apply to Intensity and
add it to a global apply group, for all fixtures, and to a global
object.