Lux: Create an ApplyGroup from a Selection
When you select some fixtures, you create a new apply group in
that action. We model this in the evaluator by pushing a new
apply group to a stack of them, and evaluating applies to the last
group.

When groups are applied to an action, they are removed from the
stack, which in this case as there is only the global action - it
removes and applies them all.