Lumen: Add Range Query Step
A query can ask for a range of fixtures ID's to add to a query,
plus some other minor improvements.

The QueryBuilder feels like additional duplication of efforts when
working on Lux, however it's important to remember that Lumen
could be used independantly of Lux, which is why it's important to
keep this more ergonomic interface.