Lumen+Candela: More robust resetting of evaluator state on reload
When we refresh the state from a change in text input, we need to
make sure that we rebuild from a clean zero state, even in the
case where we are stopped, and thing time has not moved.

There will probably be more bugs with this when we are reloading
say, when paused at 5s, and we want editing to remain at the 5s
mark. But I think most of this confusion is coming from a poorly
understood architecture, and lack of robustness in the Lux
evaluation phase. After establishing this MVP, work will shift to
making these both much more beautiful.