Lux: Sub-Queries now track their parent correctly
A sub selecting query was building, naively, from the last apply
group that had been created, but when you have multiple
independant sub-queries, or queries for that matter, they act as
if all nested one after another. Now the evaluator tracks the
parent. There has got to be a cleaner way :)