Lux: Apply statements are added to the proper Parent Apply Group
Apply statements after a sub selection where naively added to the
last sub selection created, which is not the parent. Now we always
add the evaluated apply to the parent apply group - resulting in
the correct behaviour.

This also fixes applies in the global scape not working, which can
now be correctly added to the global action with the all query. We
avoid all further queries subquerying the global query (which is
unnecessary) by not adding it to the parent list, and only
returning it when we cant find a parent.

All apply statements, regardless of location in a file, that are
global will be applied at the same time.