Candela+Lumen: Refactor connection and output control to Candela
Again, we continue with very rough cuts to spike out functionality
with the intention of coming back later to understand where the
abstractions must lie.

This commit is a major milestone as it finishes the full output
pipeline and actually can control lights via SACN :)

Candela now handles it's own connection attempts, but does not yet
handle well a connection refused state. This will come later when
I tidy this all up.