Candela: Network Status Bar
We will have many different types of events triggered by Candela
and the UI, and I thought a simple network status bar would be a
good way to get a first cut of how this might work in the future.

It turns red/green depending on if it is successful in outputting
SACN data.