Lux: Support Percentage Values for Parameters
The first step of supporting this is to differentiate between the
generators we are ...generating. At the moment we just deal with
static values but lays the ground work for parsing fades shortly.

Once we know we are parsing a static value, we simply either parse
a percentage by recognising the ending '%' or not, and then
evaluate the correct Lumen Value enum for this.