Lumen: Refactor Abstract Fixtures to have a Vec of Generators
To allows for an action to create a stack of potential generators
for each parameter of a fixture, we need to change the fixture to
have a Vec of them.

When we apply a new action, if it is the first time adding a
generator to that parameter for this action, we empty it first to
restart the stack of potential generators.

This work is to allow for a parameter to have multiple delayed
generators, that will over time evaluate to new values.