"Lux+Lumen: Test Runner"
A simple test running for the lux language, where in an example
file we have a simple format for the expected state of the
environment from that code. It makes an assumtion of running the
code at 0 time. This should be sufficient for now, but can always
expand on the syntax if needed in the future.

Some changes to Lumen are made to allow for easy inspection of the
environment for this purpose, and also to tidy up some warnings.