Lux: Preset Blocks
Unlike other lighting systems, presets are not stored in fixture
by a preset reference in the lighting engine. Instead we store a
block of lux code, to evaluate in place every time the preset is
referred too.

It may later turn out to be insuffeciently simple to handle this
functionality like this, as it requires reevaluation of all the
program code to be sure we have the correct lighting state
displayed. I don't know how much of a bottle neck this will be on
a large show, as to remain responsive we want evaluation to remain
<300ms for the whole show file.