Lux: Inline Delay Syntax
Moving delay blocks out of the top_level pool and into stmt's, we
can support those delays nested inside selections. Though this
commit does not test nested delay blocks, which probably do not
work correctly and will be implemented next, or not supported.

The following syntax

```
@1s intensity: 100 -> 0 1s
```

is syntactical sugar for

```
@1s {
  intensity: 100 -> 0 1s
}
```