Lumen: Even & Odd Query can generate fixtures and limit them
When Even/Odd is encountered as the start of the query, it adds
all even/odd ID's to the result. But any later in the query, it
discards everything but the even or odd ID's.

This may be hoplessly naive, but the query language is poorly
defined and very experimental for now.