Lux: Even/Odd Query Commands
A query command is an instruction a query to do something, with
syntax `:command`. In this case we use it to access the even and
odd selectors in lumen.

The query language is poorly defined and experimental, and one of
the bedrock features of the language. I expect it to change and
evolve rapidly. The more featureful it is the more effective Lux
becomes as a programming tool.