Lumen: Basic CurrentValue Support
A CurrentValue generator takes at the point of resolution whatever
the current value is, creates an internal Static generator of it
and then returns that. In Delay and Fade generators, we propogate
the resolve to the stored Generators.

This allows a fade to say, start from the point which the fixture
is currently at. This first cut is very basic support, and will
become more complicated when we look to retain a current value
across multiple actions. But this comes in future work :)