Lux: Basic '_' CurrentValue Syntax
The placeholder value '_' refers to the current value at the point
of resolution. The example in this commit currently does not work
as we compare the state of the abstract fixtures, rather than the
concrete resolution - which is where we fill in current values.
The fix for this is coming in another commit.