Lumen: Capture the first applied CurrentValue when applying Action
If you have a long running fade in one action, then overwrite it
with a CurrentValue in another  action, then we capture it at that
point of time. This introduces a dependancy on the patch for
action application, making it no longer purely abstract. However,
there is no other way to resolve the current value of the fade,
at least not at the moment.

If this dependancy proves to be cumbersome later down the line,
then I will introduce a future abstact resolution stage. But for
now this was the simpler approach. This commit does not test if
running back and forward in time still creates the correct states,
as history reset does not reset captured generators. But maybe
it's fine, we'll have to check :)