Lux: Evaluate Grouped Generators for Position
This is a first cut, with some rough edges that need coming back
to, but works as a first proof of concept implementation.

By decided that grouped parameters are not a Lumen concern, but
a syntactical sugar in Lux only - we can recognise certain
parameters that can be treated as a group and expand them to
their parameters.