Lumen+Lux: Make Group Parameters a static of only Lux Language
Introducing a dependancy on lazy_static, we can just keep a static
public reference to group parameters as they are understood in the
Lux language. This allows us to remove a bunch of unecessary
cloning, and simplifies the code.

It also means we can remove the artificial Param of Position,
which acted as a ghost Param. This keeps this enum restricted to
only parameters that will be translated to DMX.