Lumen: Seperate Color Resoloution from Main Parameters
Color is a special case when resolving because we may need to
convert from one colorspace to another. To do so we split the
resolution logic into to two seperate passes.

We take the color parameters and construct a Color object with
their resolved values. We, if needed, can convert that Color to
any colorspace we need. This is not yet implemented and needs
some careful consideration to avoid a NxN problem.

This is a first cut MVP with many TODO's, which will be fleshed
out with more test and features.